requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.5.0
Pillow>=9.0.0
numpy>=1.19.0
diffusers>=0.26.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0
